pyproject.toml
[project]
name = "personal-notes-powered-with-rags"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "requests>=2.33.0",
    "sentence-transformers>=5.3.0",
]
